refactor(core): the final-result builder moves out of the helpers file
The file-size ratchet is right: the carry fix grew two frozen files,
and the answer is a split, not an extension. buildFinalResult and the
failure resolution live in execution-failure.ts now — the module whose
subject they are — the round constant lives with the round-state type
it bounds, and the helpers file shrinks below its baseline (re-frozen
in the same change, as the ratchet demands).
